requirements.txt
pandas
torch
transformers
psycopg2-binary
scikit-learn